requirements.txt
numpy
Pillow
scikit-learn
tensorflow
joblib
torch
git+https://github.com/openai/CLIP.git